# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "XXC-APTD management",
    "version": "0.1.0",
    "description": "Unix socket: filesystem administrator authority. Administrative HTTP: local session, RBAC, exact Origin and CSRF for mutations. Public HTTP never serves these endpoints. Scoped project bearer tokens authorize only explicit publishing routes and suites. /admin is an equivalent private TCP API prefix."
  },
  "servers": [
    {
      "url": "https://apt.thugs.red/admin",
      "description": "Example path proxy. Confirm the administrative API origin with the operator; public /api is documentation only."
    },
    {
      "url": "https://admin.example.invalid",
      "description": "Example dedicated admin origin; replace with the operator-provided value. Paths below already include /api/v1."
    }
  ],
  "paths": {
    "/api/v1/status": {
      "get": {
        "summary": "status",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/packages": {
      "get": {
        "summary": "packages",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/uploads": {
      "get": {
        "summary": "uploads",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "post": {
        "summary": "uploads",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Package"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "408": {
            "$ref": "#/components/responses/Error408"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "application/vnd.debian.binary-package": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "x-minimum-role": "operator",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "upload"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/jobs": {
      "get": {
        "summary": "jobs",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "description": "Lookup is bounded to the latest 100 archive jobs. Project tokens see only jobs belonging to the same token. A completed job may later fall outside this window.",
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/config": {
      "get": {
        "summary": "config",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "administrator"
      }
    },
...
    "/api/v1/repository/publish": {
      "post": {
        "summary": "repository publish",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "202": {
            "description": "Queued; poll the returned job ID",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publish"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "x-minimum-role": "operator",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "publish"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/repository/diff": {
      "get": {
        "summary": "Review the current publication selection",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preview"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "summary": "HTTP only: revoke current session",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "logged_out": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "logged_out"
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-xxc-session",
        "description": "Admin HTTP interface only. When admin.external_url uses HTTPS, the cookie is __Host-xxc-session with Secure. Explicit HTTP administration uses xxc-session, independently of the public origin or proxy headers. Socket requests authenticate through filesystem permissions."
      },
      "projectToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "One-time project token with explicit scope and suite allowlists"
      }
    },
    "schemas": {
...
      "Status": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "version",
          "packages",
          "staged",
          "administrative_http",
          "generation"
        ],
        "properties": {
          "version": {
            "type": "string"
          },
          "packages": {
            "type": "integer",
            "minimum": 0
          },
          "staged": {
            "type": "integer",
            "minimum": 0
          },
          "administrative_http": {
            "type": "string",
            "enum": [
              "authentication_required",
              "disabled"
            ]
          },
          "generation": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "headers": {
      "RequestId": {
        "schema": {
          "type": "string"
        },
        "description": "Correlation identifier"
      },
      "NoStore": {
        "schema": {
          "type": "string",
          "enum": [
            "no-store"
          ]
        }
      }
    },
    "responses": {
      "Error400": {
        "description": "Invalid request or application validation failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        }
      },
      "Error404": {
        "description": "Unknown object",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ],
...
}